Se finaliza las pruebas de unidad de requerimientos

+++ SAPS/SAPS/Codigo_Fuente/Controladoras/ControladoraRequerimientos.cs
@@ -30,7 +30,7 @@ namespace SAPS.Controladoras
         }
 
         /** @brief Método que asigna las operaciones necesarias para poder insertar un requerimiento.
-         * @@param datos Un vector tipo objeto que contiene toda la información necesaria
+         * @param datos Un vector tipo objeto que contiene toda la información necesaria
                    para crear un requerimeinto, el orden de los parámetros va de la siguiente manera:
             | Índice |      Descripción     | Tipo de datos |
             |:------:|:--------------------:|:-------------:|
@@ -54,6 +54,8 @@ namespace SAPS.Controladoras
          */
         public int modificar_requerimiento(Object[] datos)
         {
+            if (!existe_requerimiento(Convert.ToInt32(datos[0]))) //Verifica que el requerimiento exista
+                return -1;
             Requerimiento requerimiento = new Requerimiento(datos);
             return m_base_datos.modificar_requerimiento(requerimiento);
         }
@@ -64,6 +66,8 @@ namespace SAPS.Controladoras
          */
         public int eliminar_requerimiento(int id_requerimiento)
         {
+            if(!existe_requerimiento(id_requerimiento))
+                return -1;
             return m_base_datos.eliminar_requerimiento(id_requerimiento);
         }
 
@@ -99,7 +103,25 @@ namespace SAPS.Controladoras
         */
         public int asociar_requerimiento(Object[] datos)
         {
-            return m_base_datos.asociar_requerimiento(datos); // Prueba
+            if(existe_requerimiento(Convert.ToInt32(datos[1]))) //Verifica que el requerimiento exista
+                return m_base_datos.asociar_requerimiento(datos);
+            return -1;
+        }
+
+        /** @brief Método que se encarga de verificar si un requerimiento existe en la base de datos.
+         * @param El identificador del requerimiento que se quiere verificar si existe.
+         * @return True si el requerimiento está en la base, False si no.
+        */
+        private bool existe_requerimiento(int id_requerimiento)
+        {
+            DataTable resultado = solicitar_requerimientos_disponibles();
+            int cant_filas = resultado.Rows.Count;
+            for(int i=0; i<cant_filas; ++i)
+            {
+                if (Convert.ToInt32(resultado.Rows[i]["id_requerimiento"]) == id_requerimiento)
+                    return true;
+            }
+            return false;
         }
     }
 }
